Infosys/agrobase/auth_backend/requirements.txt
# --- Backend Authentication (Flask) ---
Flask==3.0.3
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1
python-dotenv==1.0.1
sqlite3-binary==0.0.1  # Optional fallback for Windows
